rfctr(msg): remove temporary new_msg.py
