+++ unstructured/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.14.5-dev3"  # pragma: no cover
+__version__ = "0.14.5-dev4"  # pragma: no cover
